[frontend] refactor: split main.js God-file into screen modules (#118)
main.js held 1363 LOC: six screens, the shared state object, every Wails
binding and the boot sequence in one module. Split it along the seams the
file already had — its own `// ── Screen N ──` banners — into one module per
screen plus small shared libs, following the existing src/lib/ convention.

  main.js (146)            entry point: boot, Wails event wiring, render()
                           router, and the two boot calls
  lib/state.js (43)        the shared mutable state object
  lib/render.js (16)       the repaint indirection (see below)
  lib/branding.js (19)     applyBranding + installVerb
  lib/titlebar.js (52)     frameless-window chrome shared by every screen
  screens/launchpad.js     screen 1: form, validation, BitLocker chooser,
                           applyImageDefaults, startInstall
  screens/progress.js      screen 2 + INSTALL_STEPS
  screens/vmpreview.js     Try-in-VM preview (§6.1)
  screens/done.js          screen 3
  screens/control.js       screen 4 + confirmUninstall
  screens/migrate.js       screen 5 + refreshCategories

The crux was the coupling: every screen mutates a shared `state` and calls a
global `render()`, but render() dispatches to the screens — so importing it
from a screen would close a main → screen → main cycle. main.js keeps the
screen table and therefore owns render(); screens only ever need to *trigger*
a repaint, so main.js registers its router with lib/render.js at boot and
screens call that indirection instead. state.js and render.js import nothing,
which keeps the whole graph a DAG and rules out cycles by construction.

Pure refactor — no behaviour change. Every line of code and every comment was
moved verbatim by slicing the original at line boundaries, then checked
mechanically: all 1228 non-blank body lines of the old main.js still exist
with identical text in the new modules, and the only additions are import
statements and module headers. No DOM id, CSS class or user-facing string was
touched, so the GUI suite and the label-based E2E drive helper keep working.

The one dropped symbol is the `GetStatus` import, which was already dead.

Verified: `npm run build`, `go build ./...`, and all 19 tests in
tests/gui/gui.spec.js pass.
